# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ingestor System API
  description: |-
    RESTful API for the Ingestor System that provides access to content processing,
    entity extraction, and database capabilities using Claude AI.
  version: 1.0.0
  contact:
    email: support@example.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:3000/api/v1'
    description: Local development server
...
paths:
...
  /batch:
    get:
      tags:
        - batch
      summary: List batch processes
      description: Returns a paginated list of batch processes with optional filtering
      operationId: listBatchProcesses
      parameters:
        - name: status
          in: query
          description: Filter by process status
          schema:
            type: string
            enum: [pending, processing, completed, failed, cancelled]
        - name: limit
          in: query
          description: Maximum number of items to return
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: sort
          in: query
          description: Field to sort by
          schema:
            type: string
            default: createdAt
            enum: [createdAt, updatedAt, status]
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            default: desc
            enum: [asc, desc]
      responses:
        '200':
          description: List of batch processes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BatchProcess'
                  meta:
                    type: object
                    properties:
                      pagination:
                        type: object
                        properties:
                          limit:
                            type: integer
                            example: 20
                          offset:
                            type: integer
                            example: 0
                          total:
                            type: integer
                            example: 15
                          next:
                            type: string
                            example: null
                            nullable: true
                          previous:
                            type: string
                            example: null
                            nullable: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    
    post:
      tags:
        - batch
      summary: Create batch process
      description: Creates a new batch processing job
      operationId: createBatchProcess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchProcessCreate'
      responses:
        '202':
          description: Batch process created and queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/BatchProcess'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /batch/{id}:
    get:
      tags:
        - batch
      summary: Get batch process
      description: Returns details of a specific batch process
      operationId: getBatchProcess
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Batch process details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/BatchProcess'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Batch process not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /batch/{id}/cancel:
    post:
      tags:
        - batch
      summary: Cancel batch process
      description: Cancels a running or pending batch process
      operationId: cancelBatchProcess
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Batch process cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Batch process cancelled successfully
                      id:
                        type: string
                        format: uuid
                        example: 123e4567-e89b-12d3-a456-426614174000
                      status:
                        type: string
                        enum: [cancelled]
                        example: cancelled
        '400':
          description: Batch process cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Batch process not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /batch/{id}/items:
    get:
      tags:
        - batch
      summary: List batch process items
      description: Returns items in a batch process with their individual status
      operationId: listBatchProcessItems
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filter by item status
          schema:
            type: string
            enum: [pending, processing, completed, failed, skipped]
        - name: limit
          in: query
          description: Maximum number of items to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of batch process items
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BatchProcessItem'
                  meta:
                    type: object
                    properties:
                      pagination:
                        type: object
                        properties:
                          limit:
                            type: integer
                            example: 100
                          offset:
                            type: integer
                            example: 0
                          total:
                            type: integer
                            example: 350
                          next:
                            type: string
                            example: /api/v1/batch/123e4567-e89b-12d3-a456-426614174000/items?limit=100&offset=100
                            nullable: true
                          previous:
                            type: string
                            example: null
                            nullable: true
                      status:
                        type: object
                        properties:
                          pending:
                            type: integer
                            example: 200
                          processing:
                            type: integer
                            example: 50
                          completed:
                            type: integer
                            example: 95
                          failed:
                            type: integer
                            example: 5
                          skipped:
                            type: integer
                            example: 0
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Batch process not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  
  #
  # System Endpoints
  #
...
components:
  schemas:
...
    BatchProcessCreate:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [folder-import, url-crawl, entity-extraction, reprocess, content-analysis]
          example: folder-import
        name:
          type: string
          example: Import Customer Documentation
        description:
          type: string
          example: Batch import of customer documentation folder
        options:
          type: object
          properties:
            path:
              type: string
              example: /uploads/customer-docs/
            recursive:
              type: boolean
              example: true
            fileTypes:
              type: array
              items:
                type: string
              example: [".pdf", ".docx", ".txt"]
            extractEntities:
              type: boolean
              example: true
            urls:
              type: array
              items:
                type: string
              example: ["https://example.com/docs/", "https://example.com/guides/"]
            crawlDepth:
              type: integer
              example: 2
            contentIds:
              type: array
              items:
                type: string
                format: uuid
              example: ["123e4567-e89b-12d3-a456-426614174001", "123e4567-e89b-12d3-a456-426614174002"]
            filters:
              type: object
              properties:
                contentType:
                  type: array
                  items:
                    type: string
                  example: ["pdf", "document"]
                createdAfter:
                  type: string
                  format: date-time
                  example: '2025-04-01T00:00:00Z'
                createdBefore:
                  type: string
                  format: date-time
                  example: '2025-05-01T00:00:00Z'
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
          example: normal
        callback:
          type: object
          properties:
            url:
              type: string
              example: https://example.com/webhook/batch-complete
            headers:
              type: object
              additionalProperties:
                type: string
              example:
                X-API-Key: abcdef123456
    BatchProcess:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        name:
          type: string
          example: Import Customer Documentation
        description:
          type: string
          example: Batch import of customer documentation folder
        type:
          type: string
          enum: [folder-import, url-crawl, entity-extraction, reprocess, content-analysis]
          example: folder-import
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
          example: processing
        progress:
          type: object
          properties:
            total:
              type: integer
              example: 350
            completed:
              type: integer
              example: 95
            failed:
              type: integer
              example: 5
            processing:
              type: integer
              example: 50
            pending:
              type: integer
              example: 200
            skipped:
              type: integer
              example: 0
            percentage:
              type: number
              format: float
              example: 27.14
        options:
          type: object
          properties:
            path:
              type: string
              example: /uploads/customer-docs/
            recursive:
              type: boolean
              example: true
            fileTypes:
              type: array
              items:
                type: string
              example: [".pdf", ".docx", ".txt"]
        priority:
          type: string
          enum: [low, normal, high]
          example: normal
        createdAt:
          type: string
          format: date-time
          example: '2025-05-14T04:21:56Z'
        startedAt:
          type: string
          format: date-time
          example: '2025-05-14T04:22:00Z'
        completedAt:
          type: string
          format: date-time
          example: null
          nullable: true
        estimatedCompletionAt:
          type: string
          format: date-time
          example: '2025-05-14T05:30:00Z'
          nullable: true
        createdBy:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        links:
          type: object
          properties:
            self:
              type: string
              example: /api/v1/batch/123e4567-e89b-12d3-a456-426614174000
            items:
              type: string
              example: /api/v1/batch/123e4567-e89b-12d3-a456-426614174000/items
            cancel:
              type: string
              example: /api/v1/batch/123e4567-e89b-12d3-a456-426614174000/cancel
              nullable: true
    BatchProcessItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
        batchId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        type:
          type: string
          enum: [file, content, url]
          example: file
        status:
          type: string
          enum: [pending, processing, completed, failed, skipped]
          example: completed
        source:
          type: string
          example: /uploads/customer-docs/guide.pdf
        contentId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174002
          nullable: true
        error:
          type: object
          nullable: true
          properties:
            code:
              type: string
              example: FILE_PROCESSING_ERROR
            message:
              type: string
              example: Failed to process file due to corrupted format
        processingTime:
          type: number
          format: float
          example: 5.23
          description: Processing time in seconds
          nullable: true
        createdAt:
          type: string
          format: date-time
          example: '2025-05-14T04:22:00Z'
        startedAt:
          type: string
          format: date-time
          example: '2025-05-14T04:22:05Z'
        completedAt:
          type: string
          format: date-time
          example: '2025-05-14T04:22:10Z'
          nullable: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Validation error
            details:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                    example: content
                  message:
                    type: string
                    example: Content is required
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key